vtu/requirements.txt
absl-py==0.10.0
django-cleanup
django-cors-headers
aiohttp==3.6.2
appdirs==1.4.4
asgiref==3.3.1
astroid==1.4.9
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.1.0
blessings==1.7
bowler==0.8.0
bpython==0.19
CacheControl==0.12.6
cachetools==4.1.1
cachy==0.3.0
Cerberus==1.3.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cleo==0.7.6
click==7.1.2
clikit==0.4.3
cryptography==3.1
cs50==5.1.0
curtsies==0.3.4
cycler==0.10.0
dephell==0.7.7
dephell-archive==0.1.7
dephell-discover==0.2.10
dephell-licenses==0.1.7
dephell-links==0.1.5
dephell-markers==1.0.3
dephell-pythons==0.1.15
dephell-shells==0.1.5
dephell-specifier==0.2.2
dephell-venvs==0.1.17
dephell-versioning==0.1.2
Django==3.1.4
djangorestframework==3.12.2
docker==4.3.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
fissix==20.8.0
Flask==1.1.2
future==0.18.2
gast==0.3.3
glfw==1.12.0
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.31.0
h5py==2.10.0
html5lib==1.1
idna==2.10
isort==5.5.0
itsdangerous==1.1.0
jedi==0.13.2
jeepney==0.4.3
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
keyring==20.0.1
kiwisolver==1.2.0
lazy-object-proxy==1.5.1
lockfile==0.12.2
m2r==0.2.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
multidict==4.7.6
nltk==3.5
numpy==1.19.1
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
parso==0.8.0
pastel==0.2.0
pexpect==4.8.0
Pillow==7.2.0
pipreqs-amasad==0.4.10
pkginfo==1.5.0.1
pluggy==0.13.1
poetry==1.0.5
prompt-toolkit==3.0.7
protobuf==3.13.0
ptpython==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.4.0
pycparser==2.20
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.6.1
pylev==1.3.0
pylint==1.6.4
PyOpenGL==3.1.5
pyparsing==2.4.7
pyrsistent==0.14.11
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.21.5
pytz==2020.4
pyxel @ git+https://github.com/amasad/pyxel@8bcc2266b2fbb65f72ad686952de3838971c136e
PyYAML==5.3.1
regex==2020.7.14
replit==2.0.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
rope==0.11.0
rsa==4.6
scipy==1.4.1
SecretStorage==3.1.2
sh==1.14.0
shellingham==1.3.2
six==1.15.0
sounddevice==0.4.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow-cpu @ https://storage.googleapis.com/tensorflow/linux/cpu/tensorflow_cpu-2.2.0-cp38-cp38-manylinux2010_x86_64.whl
tensorflow-estimator==2.2.0
termcolor==1.1.0
tomlkit==0.5.11
tqdm==4.48.2
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.25.0
yarg==0.1.9
yarl==1.5.1
yaspin==1.0.0
